PCF/PCSS

+++ src/utils/tgaimage.cpp
@@ -6,7 +6,8 @@
 namespace Utils
 {
     TGAImage::TGAImage() : data_(), width_(0), height_(0), bytespp_(0) {}
-    TGAImage::TGAImage(const std::string filename) : data_(), width_(0), height_(0), bytespp_(0) {
+    TGAImage::TGAImage(const std::string filename) : data_(), width_(0), height_(0), bytespp_(0)
+    {
         read_tga_file(filename);
     }
     TGAImage::TGAImage(const int w, const int h, const int bpp) : data_(w * h * bpp, 0), width_(w), height_(h), bytespp_(bpp) {}
@@ -351,13 +352,13 @@ namespace Utils
                 }
             }
             erry += h;
-            oscanline += olinebytes;
+            oscanline += static_cast<int>(olinebytes);
             while (erry >= (int)height_)
             {
                 if (erry >= (int)height_ << 1) // it means we jump over a scanline
                     memcpy(tdata.data() + nscanline + nlinebytes, tdata.data() + nscanline, nlinebytes);
                 erry -= height_;
-                nscanline += nlinebytes;
+                nscanline += static_cast<int>(nlinebytes);
             }
         }
         data_ = tdata;
